Norm Verified Libft Functions

+++ ft_strchr.c
@@ -6,7 +6,7 @@
 /*   By: okochulo <okochulo@student.42vienna.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/13 11:53:38 by okochulo          #+#    #+#             */
-/*   Updated: 2025/05/23 12:38:21 by okochulo         ###   ########.fr       */
+/*   Updated: 2025/05/23 19:15:10 by okochulo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strchr(const char *str, int c)
 {
-	int	ct;
+	size_t	ct;
 
 	ct = 0;
 	while (str[ct])
